Add escape handling code to regex.
The consume_escape proceeduretakes a string piece that begins with '\\'
character and processes which RENodeType it represents and moves the
pointer to the end of the escape. Handles almost all of the escape
sequences that will be supported.

Several helpers are also added to test various character classes:
lower, upper, alpha, digit, word, and hex digits.

A quick struct/enum construction macro also added as a test to see if
it is useful.