Fix up a bunch of escape offsets, add proc header.
I previously thought I wanted consume_escape to leave the
pointer on the last character of the escape sequence, but I
changed my mind, so now it points to the true end of the
sequence. Changed the comments and offsets to match.

Added a thorough procedure header to consume_escape as well.
Doxygen style.